+++ examples/TodoMvc/TodoMvc.hs
@@ -1,0 +1,4 @@
+module TodoMvc (page, mainJS) where
+
+import Client (mainJS)
+import Page (page)